Track CWD (#27)
* Track CWD

Track the current working directory with the last index of each unique
command.  Update the various functions to include this information (if
available).

We choose to alter the cmdraw table with a rudimentary and partially
completed sql migration convention.  The alternative would be either to
use "cmdan" command annotations or a separate table.

For the former (cmdan) the current implementation is pretty dependent
on "prune" being the only one.. plus its awkward that it get treated
as "optional" metadata moving forward.

For the latter (separate table) this just seems like a bunch of
unnecessary joins to add everywhere to avoid doing the db migration.
HOWEVER - I may consider going with the separate table, if we decide we
want to keep the "history" of working directories the command may have
been run in...

* fix tests

* Add missing free

* Remove `cwd` from table create (#28)

---------

Co-authored-by: jkmathes <jack@jkmathes.org>